split definitions into their own test file, clean up how benches are tested, get rid of the async virus,and other refactors
