+++ @re-/assert/test/custom.snapshots.json
@@ -1,5 +1,5 @@
 {
-    "assert.test.ts": {
+    "externalSnapshots.test.ts": {
         "toCustomFile": {
             "re": "do"
         },
