+++ @re-/assert/test/metaTests/templateForBenches.ts
@@ -1,25 +1,16 @@
 import { bench } from "../../src/index.js"
 
-// prettier-ignore
-const metaBenchTests = () => {
-    //bench1
-    bench("bench call single stat", () => {
-        return "boofoozoo".includes("foo")
-    }).median()
+//median
+bench("bench call single stat", () => {
+    return "boofoozoo".includes("foo")
+}).median()
 
-    //bench2
-    bench("bench call single stat", () => {
-        return "boofoozoo".includes("foo")
-    }).mean()
+//mean
+bench("bench call single stat", () => {
+    return "boofoozoo".includes("foo")
+}).mean()
 
-    //bench3
-    bench("bench call mark", () => {
-        return /.*foo.*/.test("boofoozoo")
-    }).mark()
-
-    //bench4
-    bench("bench call mark", () => {
-        return /.*foo.*/.test("boofoozoo")
-    }).median(`121.42us`)
-}
-metaBenchTests()
+//mark
+bench("bench call mark", () => {
+    return /.*foo.*/.test("boofoozoo")
+}).mark()
