+++ @re-/assert/test/metaTests/testHelpers.ts
@@ -1,109 +1,71 @@
-import { strict } from "node:assert"
+import { copyFileSync } from "node:fs"
 import { fromHere, readJson, shell } from "@re-/node"
-import {
-    ExpressionStatement,
-    Project,
-    SourceFile,
-    SyntaxKind,
-    VariableStatement
-} from "ts-morph"
+import { ExpressionStatement, Project, SourceFile, SyntaxKind } from "ts-morph"
+
+export type TestData = {
+    statements: ExpressionStatement[]
+    expected: Record<number | string, string>
+    sourceFile: SourceFile
+    fullText: string
+    initialText: string
+}
+const snapMatch = /\.snap\((.+)?\)/
+const benchMeanMedianMatch = /\.(mean|median)\(`\d+\.\d+(ns|us|s|ms)`\)/
+export const markMatch =
+    /\.mark\(\{(mean|median): ?`\d+\.\d+(ns|us|ms|s)`, ?(mean|median): ?`\d+\.\d+(ns|us|ms|s)`\}\)/
 
 const EXPECTED_RESULTS_JSON_PATH = fromHere("expectedResults.json")
 const PATH_TO_ASSERTIONS_DIR = fromHere(".reassert")
+const testFile = "copiedTestFile.ts"
+export const testFileCopyPath = fromHere(testFile)
 
-const getExpressionStatements = (
-    functionName: string,
-    varStatements: VariableStatement[]
-) =>
-    varStatements
-        .find((statement) => statement.getText().includes(functionName))
-        ?.getDescendantsOfKind(SyntaxKind.ExpressionStatement)
+const filterExpressionStatements = (statements: ExpressionStatement[]) =>
+    statements.filter(
+        (statement) =>
+            benchMeanMedianMatch.test(statement.getText()) ||
+            snapMatch.test(statement.getText()) ||
+            markMatch.test(statement.getText())
+    )
 
-const getTestComment = (statement: ExpressionStatement, text: string) => {
+export const getTestComment = (
+    statement: ExpressionStatement,
+    fileText: string
+) => {
     const range = statement.getLeadingCommentRanges()
-    return text
+    return fileText
         .slice(range[0].getPos(), range[0].getEnd())
         .replace("//", "")
         .trim()
 }
-export const cleanUpSourceFile = async (sf: SourceFile, template: string) => {
+export const cleanUpSourceFile = (sf: SourceFile, template: string) => {
     sf.replaceWithText(template)
-    await sf.save()
-}
-export const getExpectedResult = (
-    statement: ExpressionStatement,
-    expectedData: Record<string, string>,
-    text: string
-) => expectedData[getTestComment(statement, text)]
-
-export const checkIfTestsContainErrors = (
-    statements: ExpressionStatement[],
-    expectedValues: Record<string, string>,
-    fullText: string
-) => {
-    let hasErrors = false
-    for (const statement of statements) {
-        const expected = getExpectedResult(statement, expectedValues, fullText)
-        try {
-            strict.deepEqual(statement.getText(), expected)
-        } catch (e) {
-            // @ts-ignore
-            console.log(`actual: ${e.actual} \nexpected: ${e.expected}`)
-            hasErrors = true
-            break
-        }
-    }
-    return hasErrors
-}
-export const checkThatBenchSnapGetsPopulated = (
-    statements: ExpressionStatement[]
-) => {
-    const meanMedianRegex = /\.(median|mean)\(`\d+\.\d+(ns|us|s|ms)`\)/
-    const markRegex =
-        /\.mark\(\{(mean|median): ?`\d+\.+\d+(ns|us|ms|s)\\, ?median: ?`\d+\.\d+(ns|us|ms|s)`\}\)/
-    let errors = false
-    for (const statement of statements) {
-        if (
-            !(
-                meanMedianRegex.test(statement.getText()) ||
-                markRegex.test(statement.getText())
-            )
-        ) {
-            errors = true
-            break
-        }
-    }
-    return errors
+    sf.saveSync()
 }
 
-export const getTestFileData = async (
-    filename: string,
-    templateFilename: string,
-    precache = true
-) => {
-    const filePath = fromHere(`${filename}.ts`)
+export const getTestFileData = (templateFilename: string, precache = true) => {
     const templatePath = fromHere(`${templateFilename}`)
-    const p = new Project()
-    const emptySnapsTemplateFile = p.addSourceFileAtPath(templatePath)
+    const project = new Project()
+    const emptySnapsTemplateFile = project.addSourceFileAtPath(templatePath)
     const initialText = emptySnapsTemplateFile.getFullText()
-    const sourceFile = p.addSourceFileAtPath(filePath)
-    await cleanUpSourceFile(sourceFile, initialText)
-    //run file
-    precache
-        ? shell(`pnpm ts-node ${filePath} --reassertTestPreCached`, {
-              env: { RE_ASSERT_CMD: `--cacheDir ${PATH_TO_ASSERTIONS_DIR}` }
-          })
-        : shell(`pnpm ts-node ${filePath}`, {
-              env: { RE_ASSERT_CMD: undefined }
-          })
+    copyFileSync(templatePath, testFileCopyPath)
 
-    sourceFile.refreshFromFileSystemSync()
-    const varStatements = sourceFile.getVariableStatements()
+    if (precache) {
+        shell(`pnpm ts-node ${testFileCopyPath} --reassertTestPreCached`, {
+            env: { RE_ASSERT_CMD: `--cacheDir ${PATH_TO_ASSERTIONS_DIR}` }
+        })
+    } else {
+        shell(`pnpm ts-node ${testFileCopyPath}`, {
+            env: { RE_ASSERT_CMD: undefined }
+        })
+    }
 
+    const sourceFile = project.addSourceFileAtPath(testFileCopyPath)
+    const expressionStatements = sourceFile.getDescendantsOfKind(
+        SyntaxKind.ExpressionStatement
+    )
     return {
-        statements: getExpressionStatements(filename, varStatements) || [],
+        statements: filterExpressionStatements(expressionStatements) ?? [],
         expected: readJson(EXPECTED_RESULTS_JSON_PATH),
-        fullText: sourceFile.getFullText(),
         sourceFile,
         initialText
     }
